+++ packages/sit-onyx/src/components/OnyxForm/types.ts
@@ -1,7 +1,5 @@
 import type { DensityProp } from "../../composables/density";
 import type { SkeletonProvidedProp } from "../../composables/useSkeletonState";
-import type { FormInjectedProps } from "./OnyxForm.core";
+import type { FormProps } from "./OnyxForm.core";
 
-export type OnyxFormProps = Partial<FormInjectedProps> &
-  Partial<SkeletonProvidedProp> &
-  DensityProp;
+export type OnyxFormProps = FormProps & Partial<SkeletonProvidedProp> & DensityProp;