+++ packages/sit-onyx/src/components/OnyxInput/OnyxInput.vue
@@ -2,6 +2,7 @@
 import { computed } from "vue";
 import { useDensity } from "../../composables/density";
 import { useCustomValidity } from "../../composables/useCustomValidity";
+import { useErrorClass } from "../../composables/useErrorClass";
 import { SKELETON_INJECTED_SYMBOL, useSkeletonContext } from "../../composables/useSkeletonState";
 import { FORM_INJECTED_SYMBOL, useFormContext } from "../OnyxForm/OnyxForm.core";
 import OnyxFormElement from "../OnyxFormElement/OnyxFormElement.vue";
@@ -15,9 +16,10 @@ const props = withDefaults(defineProps<OnyxInputProps>(), {
   required: false,
   autocapitalize: "sentences",
   readonly: false,
-  disabled: FORM_INJECTED_SYMBOL,
   loading: false,
   skeleton: SKELETON_INJECTED_SYMBOL,
+  disabled: FORM_INJECTED_SYMBOL,
+  showError: FORM_INJECTED_SYMBOL,
 });
 
 const emit = defineEmits<{
@@ -48,8 +50,9 @@ const patternSource = computed(() => {
   return props.pattern;
 });
 
-const { disabled } = useFormContext(props);
+const { disabled, showError } = useFormContext(props);
 const skeleton = useSkeletonContext(props);
+const errorClass = useErrorClass(showError);
 </script>
 
 <template>
@@ -58,13 +61,13 @@ const skeleton = useSkeletonContext(props);
     <OnyxSkeleton class="onyx-input-skeleton__input" />
   </div>
 
-  <div v-else :class="['onyx-input', densityClass]">
+  <div v-else :class="['onyx-input', densityClass, errorClass]">
     <OnyxFormElement v-bind="props" :error-messages="errorMessages">
       <template #default="{ id }">
         <div class="onyx-input__wrapper">
           <OnyxLoadingIndicator v-if="props.loading" class="onyx-input__loading" type="circle" />
           <input
-            :id="id"
+            :id
             v-model="value"
             v-custom-validity
             :placeholder="props.placeholder"