+++ packages/sit-onyx/src/components/OnyxSwitch/OnyxSwitch.vue
@@ -32,13 +32,14 @@ const emit = defineEmits<{
 
 const { requiredMarkerClass, requiredTypeClass } = useRequired(props);
 const { densityClass } = useDensity(props);
+
+const { disabled, showError } = useFormContext(props);
 const { vCustomValidity, errorMessages } = useCustomValidity({ props, emit });
+const shownErrorMessages = computed(() =>
+  showError.value !== false ? errorMessages.value : undefined,
+);
 
-const title = computed(() => {
-  return props.hideLabel ? props.label : undefined;
-});
-
-const { disabled } = useFormContext(props);
+const title = computed(() => (props.hideLabel && props.label) || undefined);
 const skeleton = useSkeletonContext(props);
 
 const isChecked = computed({
@@ -57,7 +58,7 @@ const isChecked = computed({
     <OnyxSkeleton v-if="!props.hideLabel" class="onyx-switch-skeleton__label" />
   </div>
 
-  <OnyxErrorTooltip v-else :disabled="disabled" :error-messages="errorMessages">
+  <OnyxErrorTooltip v-else :disabled="disabled" :error-messages="shownErrorMessages">
     <label class="onyx-switch" :class="[requiredTypeClass, densityClass]" :title="title">
       <input
         v-model="isChecked"
@@ -265,6 +266,7 @@ $input-width: calc(2 * var(--onyx-switch-icon-size) - 2 * var(--onyx-switch-cont
       font-size: 1rem;
       line-height: 1.5rem;
     }
+
     &__label {
       color: var(--onyx-color-text-icons-neutral-intense);
       font-family: var(--onyx-font-family);