+++ packages/sit-onyx/src/components/OnyxSwitch/types.ts
@@ -3,9 +3,10 @@ import type { RequiredMarkerProp } from "../../composables/required";
 import type { CustomValidityProp } from "../../composables/useCustomValidity";
 import type { SkeletonInjected } from "../../composables/useSkeletonState";
 import type { AutofocusProp, TruncationType } from "../../types";
-import type { FormInjected } from "../OnyxForm/OnyxForm.core";
+import type { FormInjectedProps } from "../OnyxForm/OnyxForm.core";
 
-export type OnyxSwitchProps = DensityProp &
+export type OnyxSwitchProps = FormInjectedProps &
+  DensityProp &
   RequiredMarkerProp &
   CustomValidityProp &
   AutofocusProp & {
@@ -17,10 +18,6 @@ export type OnyxSwitchProps = DensityProp &
      * The label of the switch.
      */
     label: string;
-    /**
-     * Whether to disable the switch and prevent user interaction.
-     */
-    disabled?: FormInjected<boolean>;
     /**
      * Shows a loading indicator.
      */