+++ packages/sit-onyx/src/components/OnyxTextarea/OnyxTextarea.vue
@@ -2,6 +2,7 @@
 import { computed } from "vue";
 import { useDensity } from "../../composables/density";
 import { useCustomValidity } from "../../composables/useCustomValidity";
+import { useErrorClass } from "../../composables/useErrorClass";
 import { SKELETON_INJECTED_SYMBOL, useSkeletonContext } from "../../composables/useSkeletonState";
 import { FORM_INJECTED_SYMBOL, useFormContext } from "../OnyxForm/OnyxForm.core";
 import OnyxFormElement from "../OnyxFormElement/OnyxFormElement.vue";
@@ -14,6 +15,7 @@ const props = withDefaults(defineProps<OnyxTextareaProps>(), {
   autocapitalize: "sentences",
   readonly: false,
   disabled: FORM_INJECTED_SYMBOL,
+  showError: FORM_INJECTED_SYMBOL,
   skeleton: SKELETON_INJECTED_SYMBOL,
   disableManualResize: false,
 });
@@ -60,8 +62,9 @@ const handleInput = (event: Event) => {
   target.parentElement?.setAttribute("data-autosize-value", target.value);
 };
 
-const { disabled } = useFormContext(props);
+const { disabled, showError } = useFormContext(props);
 const skeleton = useSkeletonContext(props);
+const errorClass = useErrorClass(showError);
 </script>
 
 <template>
@@ -74,7 +77,7 @@ const skeleton = useSkeletonContext(props);
     <OnyxSkeleton class="onyx-textarea-skeleton__input" />
   </div>
 
-  <div v-else :class="['onyx-textarea', densityClass]" :style="autosizeMinMaxStyles">
+  <div v-else :class="['onyx-textarea', errorClass, densityClass]" :style="autosizeMinMaxStyles">
     <OnyxFormElement v-bind="props" :error-messages="errorMessages">
       <template #default="{ id }">
         <div class="onyx-textarea__wrapper" :data-autosize-value="value">