+++ apps/demo-app/CHANGELOG.md
@@ -1,5 +1,20 @@
 # demo-app
 
+## 1.0.0-beta.54
+
+### Patch Changes
+
+- Updated dependencies [01f8664]
+  - sit-onyx@1.0.0-beta.54
+
+## 1.0.0-beta.53
+
+### Patch Changes
+
+- Updated dependencies [be5b415]
+  - @sit-onyx/icons@1.0.0-beta.6
+  - sit-onyx@1.0.0-beta.53
+
 ## 1.0.0-beta.52
 
 ### Patch Changes