+++ packages/vitepress-theme/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sit-onyx/vitepress-theme
 
+## 1.0.0-beta.55
+
+### Patch Changes
+
+- Updated dependencies [01f8664]
+  - sit-onyx@1.0.0-beta.54
+
+## 1.0.0-beta.54
+
+### Patch Changes
+
+- sit-onyx@1.0.0-beta.53
+
 ## 1.0.0-beta.53
 
 ### Patch Changes